+++ MiracleMilk/test/src/imu/util/i2c.h
@@ -17,4 +17,4 @@
 
 int i2c_write(unsigned char slave_addr,unsigned char reg_addr,unsigned char length,unsigned char* const data);
 
-int8_t i2c_read(unsigned char slave_addr,unsigned char reg_addr,unsigned char length,unsigned char* data);
+int i2c_read(unsigned char slave_addr,unsigned char reg_addr,unsigned char length,unsigned char* data);
